requirements.txt
# Core Framework
fastapi==0.115.11
uvicorn==0.34.0
starlette==0.46.1

# WhatsApp & Messaging
twilio==9.4.6
python-multipart==0.0.20

# AI & Machine Learning
pydantic-ai==0.0.36
pydantic-ai-slim==0.0.36
google-auth==2.38.0

# Database
pymongo==4.11.2
motor==3.7.0

# Data Validation & Models
pydantic==2.10.6
pydantic-core==2.27.2

# Environment & Configuration
python-dotenv==1.0.1

# HTTP Clients
httpx==0.28.1
requests==2.32.3
aiohttp==3.11.13

# Utilities
python-dateutil==2.9.0.post0
pyjwt==2.10.1
click==8.1.8

# Development & Debugging
colorama==0.4.6

# Optional AI Models (uncomment as needed)
# openai==1.65.5
# anthropic==0.49.0
# cohere==5.14.0
# mistralai==1.5.1
# groq==0.18.0

# Optional Database Extensions
# redis==5.0.0
# sqlalchemy==2.0.0

# Production Dependencies
# gunicorn==21.2.0
# sentry-sdk==1.40.0

# All other existing dependencies
aiohappyeyeballs==2.5.0
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.6.0
attrs==25.1.0
boto3==1.37.9
botocore==1.37.9
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
eval-type-backport==0.2.2
fastavro==1.10.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
griffe==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx-sse==0.4.0
huggingface-hub==0.29.2
idna==3.10
importlib-metadata==8.5.0
jiter==0.8.2
jmespath==1.0.1
jsonpath-python==1.0.6
logfire-api==3.7.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
opentelemetry-api==1.30.0
packaging==24.2
prompt-toolkit==3.0.50
propcache==0.3.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic-graph==0.0.36
pygments==2.19.1
pyyaml==6.0.2
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20250306
typing-extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
urllib3==2.3.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
